method-1-youtube/scripts/requirements.txt
google-api-python-client==2.142.0
pandas==2.2.2
openpyxl==3.1.5
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
tqdm==4.66.4
